shared - public graph
