+++ src/pages/dashboard/analytics/gbussiness/modalShared/shared.vue
@@ -1,0 +1,2 @@
+<template src="./template.html"></template>
+<script src="./script.js"></script>
