+++ src/pages/dashboard/analytics/gbussiness/view/graph/template.html
@@ -14,7 +14,7 @@
         </template>
         |
         <b>Shared:</b>
-        <button class="btn btn-primary btn-xs" @click="event(item)">
+        <button class="btn btn-primary btn-xs" @click="$parent.editS()">
           <i class="fa fa-share"></i> Embed
         </button>
       </div>
