+++ src/pages/dashboard/analytics/gbussiness/view/script.js
@@ -1,14 +1,33 @@
 'use strict'
 import Graphs from 'factories/graphs'
 import ViewSingle from 'mixins/view-single'
+import modalShared from '../modalShared/shared'
 
 export default {
   mixins: [ViewSingle],
+
+  components: {
+    modalShared
+  },
+
+  computed: {
+    MShared() {
+      return this.$refs.modal_shared
+    },
+  },
 
   data: function () {
     return {
       entity: Graphs,
       model: {}
+    }
+  },
+
+  methods: {
+    editS: function () {
+      this.MShared
+        .onFinishCallBack(() => this.fetchData(this.id))
+        .show(this.model)
     }
   }
 }
