+++ src/pages/dashboard/analytics/gbussiness/view/template.html
@@ -1,5 +1,7 @@
 <div>
-  
+
   <router-view></router-view>
+
+  <modal-shared class="mt15" ref="modal_shared" :entity="entity"></modal-shared>
 
 </div>
